Add CustomerViewCommand
In line with issue #12, added ability for user to view a specific `Customer`
using `CustomerViewCommand`.

Following the Law of Demeter, Model has been modified to include an
additional method `getCustomer()` which will fetch the `Customer`
corresponding to the given `id` if it exists, and place it in an
`Optional`, else an empty `Optional` is returned.